requirements_pip.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.7.0
alfred==0.3
alfred-py==2.10.6
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.7
asttokens==2.0.5
async-timeout==3.0.1
attributee==0.1.7
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
basicsr==1.4.1
beautifulsoup4==4.11.1
bidict==0.22.0
black==21.4b2
bleach==5.0.0
blinker==1.4
brotlipy==0.7.0
cachetools==5.1.0
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.5
coloredlogs==15.0.1
commonmark==0.9.1
cryptography==35.0.0
cycler==0.11.0
Cython==0.29.28
dataclasses==0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
docker-pycreds==0.4.0
dominate==2.6.0
easydict==1.9
entrypoints==0.4
executing==0.8.3
facexlib==0.2.3
faiss-gpu==1.7.2
fastjsonschema==2.16.1
flatbuffers==2.0
flit_core==3.7.1
fonttools==4.33.3
funcy==1.17
future==0.18.2
gfpgan==1.3.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.46.1
gtrending==0.3.6
h5py==3.7.0
humanfriendly==10.0
hydra-core==1.1.2
idna==3.3
imageio==2.20.0
importlib-metadata==4.11.3
importlib-resources==5.9.0
iopath==0.1.9
ipdb==0.13.9
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jpeg4py==0.1.4
jsons==1.6.3
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
llvmlite==0.38.1
lmdb==1.3.0
loguru==0.6.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mkl-service==2.4.0
mpmath==1.2.1
multidict==5.1.0
mypy-extensions==0.4.3
natsort==8.1.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.11
numba==0.55.2
numpy==1.23.2
oauthlib==3.2.0
olefile==0.46
omegaconf==2.1.2
onnx==1.11.0
onnx-simplifier==0.3.10
onnxoptimizer==0.2.7
onnxruntime==1.12.0
onnxruntime-gpu==1.11.1
opencv-python==4.5.5.64
ordered-set==4.1.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pascal-voc-writer==0.1.4
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
phx-class-registry==3.0.5
pickleshare==0.7.5
Pillow==9.2.0
pip==21.2.4
portalocker==2.5.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyLaTeX==1.4.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyu2f==0.1.5
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
require==0.1.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
Send2Trash==1.8.0
sentry-sdk==1.9.3
setproctitle==1.3.1
setuptools==61.2.0
shortuuid==1.0.9
sip==4.19.13
six==1.16.0
sklearn==0.0
smmap==3.0.5
soupsieve==2.3.2.post1
stack-data==0.3.0
sympy==1.10.1
tabulate==0.8.10
tb-nightly==2.9.0a20220401
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.15.0
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tikzplotlib==0.10.1
tinycss2==1.1.1
toml==0.10.2
torch==1.12.1+cu116
torchfile==0.1.0
torchscan==0.1.2
torchvision==0.13.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typed-ast==1.4.3
typing_extensions==4.3.0
typish==1.9.3
urllib3==1.26.11
visdom==0.1.8.9
vot-toolkit==0.5.3
vot-trax==3.0.3
wandb==0.13.1
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.6.1
wrapt==1.14.1
xdg==5.1.1
yacs==0.1.8
yapf==0.32.0
yarl==1.6.3
zipp==3.7.0
